Adding support for distributed classes

+++ lyncs_mpi/client.py
@@ -1,6 +1,5 @@
 """
-Utils for interfacing dask and MPI.
-Most of the content of this file is explained in the [notebook](notebooks/Dask-mpi.ipynb).
+Specialized dask Client for MPI communicators
 """
 
 __all__ = [
@@ -16,32 +15,12 @@ import atexit
 import tempfile
 import multiprocessing
 from functools import wraps
-
-import sys
-from packaging import version
 import sh
-
-# The following can be deleted when sh==1.13.2 has been released
-if version.parse(sh.__version__) < version.parse("1.13.2"):
-
-    sys.meta_path.pop(0)
-
-    def find_spec(self, fullname, path=None, target=None):
-        """ find_module() is deprecated since Python 3.4 in favor of find_spec() """
-
-        from importlib.machinery import ModuleSpec
-
-        found = self.find_module(fullname, path)
-        return ModuleSpec(fullname, found) if found is not None else None
-
-    sh._SelfWrapper__self_module.ModuleImporterFromVariables.find_spec = find_spec
-    sh._SelfWrapper__self_module.register_importer()
-
-
 from dask_mpi import initialize
 from dask.distributed import Client as _Client
 from dask.distributed import default_client as _default_client
 from .lib import default_comm
+from .comm import Comm
 
 
 @wraps(_default_client)
@@ -53,7 +32,11 @@ def default_client():
 
 
 class Client(_Client):
-    "Wrapper to dask.distributed.Client"
+    """
+    Subclass of dask.distributed.Client specialized for MPI communicators
+    The initialization follows the guidelines of http://mpi.dask.org/
+    automatizing the process of creating MPI-distributed dask workers.
+    """
 
     def __init__(self, num_workers=None, threads_per_worker=1, launch=None):
         """
@@ -116,7 +99,7 @@ class Client(_Client):
 
         # Waiting for all the workers to connect
         def handler(signum, frame):
-            if self._server is not None:
+            if self.server is not None:
                 self.close_server()
             raise RuntimeError(
                 "Couldn't connect to %d processes. Got %d workers."
@@ -144,14 +127,20 @@ class Client(_Client):
         "Returns the list of workers."
         return self.scheduler_info()["workers"]
 
+    @property
+    def server(self):
+        "Returns the running server if available"
+        return self._server
+
     def close_server(self):
         """
         Closes the running server
         """
-        assert self._server is not None
+        if self.server is None:
+            raise RuntimeError("No MPI-server started by the client")
         self.shutdown()
         self.close()
-        self._server.wait()
+        self.server.wait()
         shutil.rmtree(self._dir)
         self._server = None
         atexit.unregister(self.close_server)
@@ -160,7 +149,7 @@ class Client(_Client):
         """
         In case of server started, closes the server
         """
-        if self._server is not None:
+        if self.server is not None:
             self.close_server()
 
         if hasattr(_Client, "__del__"):
@@ -229,19 +218,18 @@ class Client(_Client):
 
         return selected
 
-    def create_comm(self, actor=False, **kwargs):
+    @wraps(select_workers)
+    def create_comm(self, *args, **kwargs):
         """
         Return a MPI communicator involving workers available by the client.
 
         Parameters
         ----------
-        actor: bool, default True
-            Wether the returned communicator should be a dask actor.
-        **kwargs: params
+        *args, **kwargs: params
             Following list of parameters for the function select_workers.
         """
 
-        workers = self.select_workers(**kwargs)
+        workers = self.select_workers(*args, **kwargs)
         ranks = [[self.ranks[w] for w in workers]] * len(workers)
         ranks = self.scatter(ranks, workers=workers, hash=False, broadcast=False)
 
@@ -262,123 +250,4 @@ class Client(_Client):
             comm = default_comm()
             return comm.Create_group(comm.group.Incl(ranks))
 
-        return Comm(self.map(_create_comm, ranks, actor=actor))
-
-
-class Comm:
-    "MPI communicator"
-
-    __slots__ = [
-        "_comms",
-        "_ranks",
-        "_workers",
-    ]
-
-    def __init__(self, comms):
-        self._comms = tuple(comms)
-        self._ranks = self.client.map(lambda comm: comm.rank, self)
-        self._ranks = tuple(_.result() for _ in self._ranks)
-        self._workers = tuple(map(next, map(iter, map(self.client.who_has, self))))
-        assert len(set(self.workers)) == len(self), "Not unique set of workers"
-
-    @property
-    def client(self):
-        "Client of the communicator"
-        return self._comms[0].client
-
-    @property
-    def size(self):
-        "Size of the communicator"
-        return len(self)
-
-    @property
-    def ranks(self):
-        "Ranks of the communicator with respective worker"
-        return self._ranks
-
-    @property
-    def workers(self):
-        "Workers of the communicator with respective rank"
-        return self._workers
-
-    @property
-    def ranks_worker(self):
-        "Mapping between ranks and workers"
-        return dict(zip(self._ranks, self._workers))
-
-    def create_cart(self, dims, periods=None, reorder=False):
-        """
-        Makes a new communicator to which topology information has been attached
-
-        Parameters
-        ----------
-        dims: list
-            integer array specifying the number of processes in each dimension
-        periods: boolean [list]
-            logical [array] specifying whether the grid is periodic (True) or not (False)
-        reorder: boolean
-            ranking may be reordered (True) or not (False)
-        """
-        return Cartcomm(
-            self.client.map(
-                lambda comm: comm.Create_cart(dims, periods=periods, reorder=reorder),
-                self,
-            )
-        )
-
-    def __getitem__(self, key):
-        if isinstance(key, int) and key in self.ranks:
-            return self._comms[self.ranks.index(key)]
-        if isinstance(key, str) and key in self.workers:
-            return self._comms[self.workers.index(key)]
-        raise KeyError(f"{key} is neither a rank or a worker of {self}")
-
-    def __len__(self):
-        return len(self._comms)
-
-    def __iter__(self):
-        return iter(self._comms)
-
-
-class Cartcomm(Comm):
-    "Cartesian communicator"
-
-    __slots__ = [
-        "_dims",
-        "_periods",
-        "_coords",
-    ]
-
-    def __init__(self, comms):
-        super().__init__(comms)
-        topos = self.client.map(lambda comm: comm.Get_topo(), self)
-        topos = tuple(_.result() for _ in topos)
-        self._dims = tuple(topos[0][0])
-        self._periods = tuple(topos[0][1])
-        self._periods = tuple(bool(_) for _ in self._periods)
-        self._coords = tuple(tuple(topo[2]) for topo in topos)
-
-    @property
-    def dims(self):
-        "Dimensions of the cartesian communicator"
-        return self._dims
-
-    @property
-    def periods(self):
-        "Periodicity of the cartesian communicator"
-        return self._periods
-
-    @property
-    def coords(self):
-        "Coordinates of the cartesian communicator"
-        return self._coords
-
-    @property
-    def ranks_coords(self):
-        "Coordinates of the ranks of the cartesian communicator"
-        return dict(zip(self.ranks, self.coords))
-
-    def __getitem__(self, key):
-        if isinstance(key, tuple) and key in self.coords:
-            return self._comms[self.coords.index(key)]
-        return super().__getitem__(key)
+        return Comm(self.map(_create_comm, ranks))
